fix: explore swipe
- refactor: categories horizontal caips
- refactor: quiz vertical swipe
