+++ src/pages/explore/explore-page.tsx
@@ -1,24 +1,18 @@
-import { useEffect, useRef, useState } from 'react'
 import Marquee from 'react-fast-marquee'
-
-import { motion } from 'framer-motion'
-import SwiperCore from 'swiper'
-import { Mousewheel } from 'swiper/modules'
-import { Swiper, SwiperSlide } from 'swiper/react'
 
 import { withHOC } from '@/app/hoc/with-page-config'
 import HeaderOffsetLayout from '@/app/layout/header-offset-layout'
 
-import { Category } from '@/entities/category/api'
+import CategoriesHorizontalChips from '@/features/explore/ui/categories-horizontal-chips'
+import QuizVerticalSwipe from '@/features/explore/ui/quiz-vertical-swipe'
+
 import { useGetCategories } from '@/entities/category/api/hooks'
-import { GetAllQuizzesDto } from '@/entities/quiz/api'
 
 import { IcChevronRight, IcLibrary, IcLogo, IcProfile, IcSearch } from '@/shared/assets/icon'
-import { ExploreQuizCard } from '@/shared/components/cards/explore-quiz-card'
 import { Header } from '@/shared/components/header'
-import { Chip } from '@/shared/components/ui/chip'
+import QuestionBox from '@/shared/components/items/question-box-item'
 import { Text } from '@/shared/components/ui/text'
-import { Link, useQueryParam } from '@/shared/lib/router'
+import { Link } from '@/shared/lib/router'
 import { cn } from '@/shared/lib/utils'
 
 const exampleQuestions = [
@@ -28,33 +22,6 @@ const exampleQuestions = [
   { emoji: '👠 ', question: '프로세스는 무엇인가요?' },
   { emoji: '🪶', question: '숏 전략은 매수하는 전략이다' },
 ]
-
-const quizzes = [
-  {
-    id: 0,
-    name: 'picktoss',
-    question: '데킬라의 주 원료는 멕시코 할리스코 주에 서식하는 옥수수인가요?',
-    answer: 'correct',
-    explanation: '데킬라의 주 원료는 멕시코 할리스코 주에 서식하는 옥수수',
-    quizType: 'MIX_UP',
-  },
-  {
-    id: 1,
-    name: 'picktoss',
-    question: '데킬라의 주 원료는 멕시코 할리스코 주에 서식하는 옥수수인가요?',
-    answer: 'correct',
-    explanation: '데킬라의 주 원료는 멕시코 할리스코 주에 서식하는 옥수수',
-    quizType: 'MIX_UP',
-  },
-  {
-    id: 2,
-    name: 'picktoss',
-    question: '데킬라의 주 원료는 멕시코 할리스코 주에 서식하는 옥수수인가요?',
-    answer: 'correct',
-    explanation: '데킬라의 주 원료는 멕시코 할리스코 주에 서식하는 옥수수',
-    quizType: 'MIX_UP',
-  },
-] as GetAllQuizzesDto[]
 
 const ExplorePage = () => {
   const { data } = useGetCategories()
@@ -109,7 +76,7 @@ const ExplorePage = () => {
             실시간 퀴즈
           </Text>
 
-          <ScrollableChips categories={data} />
+          <CategoriesHorizontalChips categories={data} />
 
           <button
             type="button"
@@ -130,10 +97,7 @@ const ExplorePage = () => {
             <IcChevronRight className="size-[16px] text-icon-secondary" />
           </button>
 
-          <VerticalSwipeList />
-          {/* <div className="sticky top-[calc(var(--header-height-safe)+46px)] w-full h-[calc(100vh-184px)] p-[16px] pt-[48px] flex flex-col items-center gap-[10px] overflow-hidden">
-            <VerticalSwipeList />
-          </div> */}
+          <QuizVerticalSwipe />
         </div>
       </HeaderOffsetLayout>
     </>
@@ -145,205 +109,3 @@ export default withHOC(ExplorePage, {
   navClassName: 'border-t border-divider',
   backgroundClassName: 'bg-surface-2 h-fit',
 })
-
-function QuestionBox({
-  emoji,
-  question,
-  className,
-}: {
-  emoji: string
-  question: string
-  className?: HTMLElement['className']
-}) {
-  return (
-    <div
-      className={cn('px-2.5 py-1.5 bg-base-1 rounded-lg inline-flex justify-center items-center gap-2.5', className)}
-    >
-      <div className="flex items-center gap-1">
-        <Text typo="body-2-medium" color="secondary" className="leading-none">
-          {emoji} {question}
-        </Text>
-      </div>
-    </div>
-  )
-}
-
-function ScrollableChips({ categories }: { categories?: Category[] }) {
-  const [params, setParams] = useQueryParam('/explore')
-  const activeCategory = params.category
-
-  const setCategory = (categoryId: number) => {
-    setParams({ ...params, category: categoryId })
-  }
-
-  const scrollRef = useRef<HTMLDivElement>(null)
-
-  useEffect(() => {
-    const el = scrollRef.current
-    if (!el) return
-
-    const handleWheel = (e: WheelEvent) => {
-      if (e.deltaY !== 0) {
-        e.preventDefault() // ✅ preventDefault 이제 가능
-        el.scrollLeft += e.deltaY
-      }
-    }
-
-    el.addEventListener('wheel', handleWheel, { passive: false }) // ✅ 핵심
-
-    return () => {
-      el.removeEventListener('wheel', handleWheel)
-    }
-  }, [])
-
-  return (
-    <div
-      ref={scrollRef}
-      className="sticky top-[var(--header-height-safe)] bg-base-2 flex gap-[6px] overflow-x-auto scrollbar-hide px-[8px] py-[8px]"
-    >
-      {/* 전체 */}
-      <Chip
-        variant={activeCategory === 0 ? 'selected' : 'darken'}
-        left={activeCategory === 0 ? '💫' : undefined}
-        onClick={() => setCategory(0)}
-        className={cn('ml-[16px]')}
-      >
-        전체
-      </Chip>
-
-      {/* Chip 요소들 */}
-      {categories &&
-        categories.map((category, index) => (
-          <Chip
-            key={index}
-            variant={category.id === activeCategory ? 'selected' : 'darken'}
-            left={category.id === activeCategory ? category.emoji : undefined}
-            onClick={() => setCategory(category.id)}
-          >
-            {category.name}
-          </Chip>
-        ))}
-    </div>
-  )
-}
-
-function VerticalSwipeList() {
-  const [activeIndex, setActiveIndex] = useState(0)
-  const swiperRef = useRef<SwiperCore>(null)
-  const startYRef = useRef<number>(0)
-
-  useEffect(() => {
-    const handleTouchStart = (event: TouchEvent) => {
-      if (event.touches.length > 0) {
-        startYRef.current = event.touches[0].clientY
-      }
-    }
-
-    const handleTouchMove = (event: TouchEvent) => {
-      if (startYRef.current !== null && event.touches.length > 0) {
-        const currentY = event.touches[0].clientY
-        const scrollingUp = currentY < startYRef.current
-        const swiperTopOffset = swiperRef.current?.el?.getBoundingClientRect().top ?? 0
-        updateSwiperLock(swiperTopOffset, scrollingUp)
-      }
-    }
-
-    const handleWheel = (event: WheelEvent) => {
-      const swiperTopOffset = swiperRef.current?.el?.getBoundingClientRect().top ?? 0
-      const scrollingUp = event.deltaY < 0
-      updateSwiperLock(swiperTopOffset, scrollingUp)
-    }
-
-    window.addEventListener('wheel', handleWheel, { passive: false })
-    window.addEventListener('touchstart', handleTouchStart, { passive: false })
-    window.addEventListener('touchmove', handleTouchMove, { passive: false })
-
-    return () => {
-      window.removeEventListener('wheel', handleWheel)
-      window.removeEventListener('touchstart', handleTouchStart)
-      window.removeEventListener('touchmove', handleTouchMove)
-    }
-  }, [activeIndex])
-
-  const updateSwiperLock = (swiperTopOffset: number, scrollingUp: boolean) => {
-    if (!swiperRef.current) return
-
-    const safeAreaInsetTop = parseInt(
-      getComputedStyle(document.documentElement).getPropertyValue('--safe-area-inset-top') || '0',
-    )
-    const isBeyond = swiperTopOffset > 150 + safeAreaInsetTop
-    const isAtTop = activeIndex === 0
-
-    if (isBeyond) {
-      swiperRef.current.mousewheel.disable()
-      swiperRef.current.allowTouchMove = false
-    } else {
-      if (isAtTop && scrollingUp) {
-        swiperRef.current.mousewheel.disable()
-        swiperRef.current.allowTouchMove = false
-      } else {
-        swiperRef.current.mousewheel.enable()
-        swiperRef.current.allowTouchMove = true
-      }
-    }
-  }
-
-  return (
-    <motion.div
-      onPan={(_, info) => {
-        const swiperTopOffset = swiperRef.current?.el?.getBoundingClientRect().top ?? 0
-        const scrollingUp = info.delta.y < 0
-        updateSwiperLock(swiperTopOffset, scrollingUp)
-      }}
-      className="sticky top-[calc(var(--header-height-safe)+46px)] w-full h-[calc(100vh-184px)] p-[16px] pt-[48px] flex flex-col items-center gap-[10px] overflow-hidden"
-    >
-      <Swiper
-        direction="vertical"
-        slidesPerView={1}
-        spaceBetween={0.01}
-        mousewheel={{
-          forceToAxis: true,
-          enabled: false, // 초기에 비활성화
-        }}
-        modules={[Mousewheel]}
-        onSwiper={(swiper) => (swiperRef.current = swiper)}
-        onSlideChange={(swiper) => setActiveIndex(swiper.activeIndex)}
-        style={{ height: '500px', width: '100%', display: 'flex', justifyContent: 'center' }}
-      >
-        {Array.from({ length: 3 }).map((_, index) => (
-          <SwiperSlide key={index}>
-            <ExploreQuizCard
-              index={index}
-              activeIndex={activeIndex}
-              header={
-                <ExploreQuizCard.Header
-                  owner={'picktoss'}
-                  isBookmarked={false}
-                  onClickShare={() => {}}
-                  onClickBookmark={() => {}}
-                />
-              }
-              content={
-                <ExploreQuizCard.Content
-                  emoji={'🪶'}
-                  title={'인지주의 심리학 관련 퀴즈 모음'}
-                  category={'IT·개발'}
-                  playedCount={345}
-                  bookmarkCount={28}
-                />
-              }
-              quizzes={
-                <ExploreQuizCard.Quizzes
-                  quizzes={quizzes}
-                  totalQuizCount={quizzes.length}
-                  onClickViewAllBtn={() => {}}
-                />
-              }
-              footer={<ExploreQuizCard.Footer onClickStartQuiz={() => {}} />}
-            />
-          </SwiperSlide>
-        ))}
-      </Swiper>
-    </motion.div>
-  )
-}
